Add newline at the end of PreferencesWindow.cpp
This completes the move of clock preferences from Deskbar to Time.

This closes #7331. Also closing #8769 as invalid as it has been
discussed to death and 12/24 hour setting seems to belong in Locale
prefs only.

Clock preferences have been moved into their own settings file as
well so perhaps will be moved to their own replicant in the future.