Launch ReadOnlyBootPrompt if $freshInstallIndicator exists.
That file will exist on every read-only media, first boot of writable images
(usb,vmdk), and (if installing directly to partition via the build system or a
mounted image file) hard disk.

The last scenario (of installing directly to hard disk through the build system
or mounted image file) may not be ideal, but it generally tends to be an
advanced installation method. Additionally, it will occur only on first boot.

Fixes #8945.