Fix a bug involving the Vulcan Death Grip closing the wrong app
If you have expander turned on with expanded apps and you quickly remove teams with the VDG
you can remove a team not under your mouse pointer, instead you remote the team above. This
is because the window watcher thread hasn't updated yet so the TeamItemAtPoint() method
reads a window menu item instead of the team item. The solution is to lock the window
watcher thread and explicitly remove the window menu items in RemoveTeam().

This bug can be really bad if you accidentially VDG Tracker as your system gets hosed until
you restart Tracker or reboot.