Fix "Error opening terminal:xterm-256color" issue (#9636)
Looks like switching to declare "xterm-256color" terminal emulation was
made a bit early: there are lot of servers that still do not know about
this terminal. As was discussed in #9636 the only acceptable way is to
switch back to "xterm" and adjust corresponding entry in our local
termcap database to support 256 colors. So this changeset:

* Declare emulated terminal as "xterm";
* Change the colors and color pairs of "xterm" termcap entry to support
  256 colors;

Workarounds the #9636. Should be upgraded to "xterm-256color" some time
in the future.