Terminal: Move secondary windows back in screen
For the Window title and tab title edit windows there was already code used to
move a window that had gone out of the screen frame back in.

I generalized this code by turning it into a _MoveWindowInScreen() method
and then called it in 3 places, the original 2 cases as well as the Find window.
We might want to move this method into BWindow if this is something we'd like
to use it for windows in other applications, but this solves the problem in Terminal
for now.