Backgrounds: Line up BBox top borders refactor
Found an AddStrut() method that eliminates the need for the ugly
CreateVerticalStut() call. This approximately matches what I did
in Screen Preferences to line up the BBox's there.

I've reverted my previous commit and redid the code to make
the history as nice as possible but my main concern is to make the
code as nice as possible.