Mail: Tweak the height of the menu fields
Since hrev45725 we no longer resize fixed sized menu fields to their preferred
height in the constructor so as to be backwards compatible with the behavior on
BeOS R5. As a consequence though, the menu fields in Mail are now a bit too
tall so this commit tweaks the menu fields to be just a bit shorter matching
the height of the text controls.